moved book location to new component
