+++ src/app/components/book-list/book-detail.modal.ts
@@ -2,7 +2,6 @@ import { Component, OnInit, Input } from '@angular/core';
 import { ModalController } from '@ionic/angular';
 import { HttpHeaders, HttpParams, HttpClient } from '@angular/common/http';
 import { CacheService } from 'ionic-cache';
-import * as moment from 'moment';
 import { TranslateService } from '@ngx-translate/core';
 import { AlertService } from '../../services/alert/alert.service';
 import { IConfig } from '../../lib/interfaces';
@@ -24,13 +23,12 @@ export class BookDetailModalPage implements OnInit {
   showShortAbstract = true;
   showFullTOC = false;
   isLoaded = false;
+  locationData;
 
   @Input() book;
   @Input() isFavorite;
-  bookLocationList = [];
   shortAbstract = false;
   bookDetails = {
-    'url': null,
     'keywords': [],
     'isbn': [],
     'series': [],
@@ -111,7 +109,11 @@ export class BookDetailModalPage implements OnInit {
       if (refresher) {
         refresher.target.complete();
       }
-      this.setLocationData(data);
+
+      if (data) {
+        this.locationData = data;
+      }
+
       this.isLoaded = true;
     }, error => {
       console.log(error);
@@ -120,185 +122,6 @@ export class BookDetailModalPage implements OnInit {
         refresher.target.complete();
       }
     });
-  }
-
-  /**
-   * @name setLocationData
-   * @param data
-   */
-  setLocationData(data) {
-    // console.log(data);
-    this.bookLocationList = [];
-    if (data && data.document) {
-      let i, j;
-      data.document = utils.convertToArray(data.document);
-      for (i = 0; i < data.document.length; i++) {
-        if (data.document[i].item) {
-          data.document[i].item = utils.convertToArray(data.document[i].item);
-          for (j = 0; j < data.document[i].item.length; j ++) {
-            const locationModel = {
-              department:     this.getDepartment(data.document[i].item[j]),
-              departmentURL:  this.getDepartmentURL(data.document[i].item[j]),
-              label:          this.getLabel(data.document[i].item[j]),
-              item:           this.getItem(data.document[i].item[j]),
-              url:            this.getBookUrl(data.document[i].item[j])
-            };
-            this.bookLocationList.push(locationModel);
-          }
-        }
-      }
-    }
-    // console.log(this.bookLocationList);
-  }
-
-  /**
-   * @name getDepartment
-   * @param item
-   */
-  getDepartment(item) {
-    let department = '';
-    if (item.department && item.department.content) {
-      department = item.department.content;
-    }
-    if (item.storage) {
-      department = department + ', ' + item.storage.content;
-    }
-    return department;
-  }
-
-  /**
-   * @name getDepartmentURL
-   * @param item
-   */
-  getDepartmentURL(item) {
-    if (item.department && item.department.id) {
-      return item.department.id;
-    } else { return ''; }
-  }
-
-  /**
-   * @name getLabel
-   * @param item
-   */
-  getLabel(item) {
-    if (item.label) {
-      return item.label;
-    } else { return ''; }
-  }
-
-  /**
-   * @name getBookUrl
-   * @param item
-   */
-  getBookUrl(item): void {
-    if (this.book.location) {
-      let i;
-      const tmp = utils.convertToArray(this.book.location);
-      for (i = 0; i < tmp.length; i++) {
-        if (tmp[i].url) {
-          const tmpUrl = utils.convertToArray(tmp[i].url);
-          let j;
-          for (j = 0; j < tmpUrl.length; j++) {
-            if (tmpUrl[j].$ && (tmpUrl[j].$.usage === 'primary display')) {
-              if (tmpUrl[j]._) {
-                this.bookDetails.url = tmpUrl[j]._;
-              }
-            }
-          }
-        }
-      }
-    }
-
-    if (this.bookDetails.mediaType === 'mediatype_o') {
-      let url;
-      if (item.unavailable && item.unavailable[0].service === 'openaccess') {
-        url = item.unavailable[0].href;
-      } else { url = null; }
-      if (url != null) { this.bookDetails.url = url; }
-    }
-
-    return this.bookDetails.url;
-  }
-
-  /**
-   * @name getItem
-   * @param item
-   */
-  getItem(item) {
-    let status = '', statusInfo = '';
-
-    // check for available / unavailable items and process loan and presentation
-    let loanAvailable, presentationAvailable;
-    if (item.available) {
-      // tslint:disable-next-line:no-var-keyword
-      const availableArray = utils.convertToArray(item.available);
-      loanAvailable = availableArray.find(x => x.service === 'loan');
-      presentationAvailable = availableArray.find(x => x.service === 'presentation');
-    }
-
-    let loanUnavailable, presentationUnavailable;
-    if (item.unavailable) {
-      const unavailableArray = utils.convertToArray(item.unavailable);
-      loanUnavailable = unavailableArray.find(x => x.service === 'loan');
-      presentationUnavailable = unavailableArray.find(x => x.service === 'presentation');
-    }
-
-    if (loanAvailable) {
-      status = 'ausleihbar';
-
-      if (presentationAvailable) {
-        // tag available with service="loan" and href=""?
-        if (presentationAvailable.limitation) {
-          statusInfo = presentationAvailable.limitation[0].content;
-        }
-
-        if (loanAvailable.href === '') {
-          statusInfo = statusInfo + 'Bitte bestellen';
-        }
-      }
-    } else {
-      // check for loan in unavailable items
-      // indicates LBS and Online-Resources
-      if (loanUnavailable && loanUnavailable.href) {
-        if (loanUnavailable.href.indexOf('loan/RES') !== -1) {
-          status = 'ausleihbar';
-        } else { status = 'nicht ausleihbar'; }
-      } else {
-        // if there is no url then it will be a presentation
-        if (this.getBookUrl(item) == null) {
-          if (item.label && item.label.indexOf('bestellt') !== -1) {
-            status = item.label;
-            statusInfo = '';
-          } else {
-            status = 'Präsenzbestand';
-            if (presentationAvailable.limitation) {
-              statusInfo = presentationAvailable.limitation[0].content;
-            }
-          }
-        } else {
-          status = 'Online-Ressource im Browser öffnen';
-        }
-      }
-
-      if (presentationUnavailable) {
-        if (loanUnavailable && loanUnavailable.href) {
-          if (loanUnavailable.href.indexOf('loan/RES') !== -1) {
-            status = 'ausgeliehen';
-            if (!loanUnavailable.expected || loanUnavailable.expected === 'unknown') {
-              statusInfo = statusInfo + 'ausgeliehen, Vormerken möglich';
-            } else {
-              statusInfo = statusInfo + 'ausgeliehen bis ';
-              statusInfo = statusInfo + moment(loanUnavailable.expected, 'YYYY-MM-DD').format('DD.MM.YYYY');
-              statusInfo = statusInfo + ', vormerken möglich';
-            }
-          }
-        } else {
-          statusInfo = statusInfo + '...';
-        }
-      }
-    }
-
-    return [status, statusInfo];
   }
 
   /**
