+++ src/app/components/book-location/book-location.component.html
@@ -1,0 +1,25 @@
+<ion-list>
+  <ion-item style="--padding-start: 5px;" lines="none" class="department" color="primary">
+      <ion-label text-wrap>
+          {{ departmentName }}
+      </ion-label>
+      <ion-icon slot="end" style="color: white; -webkit-margin-start: 0;" name="information-circle" (click)="webIntent.permissionPromptWebsite(departmentURL)"></ion-icon>
+  </ion-item>
+  <ion-item style="--padding-start: 5px;" lines="none" *ngIf="url; else noUrl" color="secondary" class="itemStatus">
+      <ion-label text-wrap>
+          <span (click)="webIntent.permissionPromptWebsite(url)" style="text-decoration: underline; color: var(--ion-color-primary)">{{ item[0] }}</span>
+      </ion-label>
+  </ion-item>
+  <ng-template #noUrl>
+      <ion-item style="--padding-start: 5px;" lines="none" color="secondary" class="itemStatus">
+          <ion-icon slot="start" style="color: darkred; -webkit-margin-end: 10px;" *ngIf="item[0] && (item[0].indexOf('ausgeliehen') != -1) || (item[0].indexOf('nicht') != -1) || (item[0].indexOf('bestellt') != -1)" name="ios-remove-circle"></ion-icon>
+          <ion-icon slot="start" style="color: darkgreen; -webkit-margin-end: 10px;" *ngIf="item[0] && departmentName && (item[0].indexOf('Präsenzbestand') != -1) || (departmentName.indexOf('Handapparat') != -1)" name="ios-contrast"></ion-icon>
+          <ion-icon slot="start" style="color: darkgreen; -webkit-margin-end: 10px;" *ngIf="!(item[0] && ((item[0].indexOf('ausgeliehen') != -1) || (item[0].indexOf('nicht') != -1) || (item[0].indexOf('bestellt') != -1)) || item[0] && departmentName && ((item[0].indexOf('Präsenzbestand') != -1) || (departmentName.indexOf('Handapparat') != -1)))" name="ios-checkmark-circle"></ion-icon>
+          <ion-label text-wrap style="color: var(--ion-color-primary);">
+              {{ item[0] }}
+              <span *ngIf="item[1]" style="color: black; font-size: small"><br>{{ item[1] }}</span>
+          </ion-label>
+          <ion-note *ngIf="label && item[0] && label != item[0]" style="font-size: small" slot="end">{{ label }}</ion-note>
+      </ion-item>
+  </ng-template>
+</ion-list>
