+++ src/app/components/book-location/book-location.component.ts
@@ -1,0 +1,187 @@
+import { Component, OnInit, Input } from '@angular/core';
+import * as moment from 'moment';
+import { utils } from '../../lib/util';
+import { WebIntentService } from '../../services/web-intent/web-intent.service';
+
+@Component({
+  selector: 'app-book-location',
+  templateUrl: './book-location.component.html',
+  styleUrls: ['./book-location.component.scss']
+})
+export class BookLocationComponent implements OnInit {
+
+  @Input() bookLocation;
+  @Input() mediaType;
+  @Input() department;
+
+  departmentName;
+  departmentURL;
+  label;
+  item;
+  url;
+
+  constructor(
+    public webIntent: WebIntentService // is used in the HTML
+  ) { }
+
+  ngOnInit() {
+    this.departmentName = this.getDepartment(this.department);
+    this.departmentURL = this.getDepartmentURL(this.department);
+    this.label = this.getLabel(this.department);
+    this.item = this. getItem(this.department);
+    this.url = this.getBookUrl(this.department);
+  }
+
+  /**
+   * @name getDepartment
+   * @param item
+   */
+  getDepartment(item) {
+    let department = '';
+    if (item.department && item.department.content) {
+      department = item.department.content;
+    }
+    if (item.storage) {
+      department = department + ', ' + item.storage.content;
+    }
+    return department;
+  }
+
+  /**
+   * @name getDepartmentURL
+   * @param item
+   */
+  getDepartmentURL(item) {
+    if (item.department && item.department.id) {
+      return item.department.id;
+    } else { return ''; }
+  }
+
+  /**
+   * @name getLabel
+   * @param item
+   */
+  getLabel(item) {
+    if (item.label) {
+      return item.label;
+    } else { return ''; }
+  }
+
+  /**
+   * @name getItem
+   * @param item
+   */
+  getItem(item) {
+    let status = '', statusInfo = '';
+
+    // check for available / unavailable items and process loan and presentation
+    let loanAvailable, presentationAvailable;
+    if (item.available) {
+      // tslint:disable-next-line:no-var-keyword
+      const availableArray = utils.convertToArray(item.available);
+      loanAvailable = availableArray.find(x => x.service === 'loan');
+      presentationAvailable = availableArray.find(x => x.service === 'presentation');
+    }
+
+    let loanUnavailable, presentationUnavailable;
+    if (item.unavailable) {
+      const unavailableArray = utils.convertToArray(item.unavailable);
+      loanUnavailable = unavailableArray.find(x => x.service === 'loan');
+      presentationUnavailable = unavailableArray.find(x => x.service === 'presentation');
+    }
+
+    if (loanAvailable) {
+      status = 'ausleihbar';
+
+      if (presentationAvailable) {
+        // tag available with service="loan" and href=""?
+        if (presentationAvailable.limitation) {
+          statusInfo = presentationAvailable.limitation[0].content;
+        }
+
+        if (loanAvailable.href === '') {
+          statusInfo = statusInfo + 'Bitte bestellen';
+        }
+      }
+    } else {
+      // check for loan in unavailable items
+      // indicates LBS and Online-Resources
+      if (loanUnavailable && loanUnavailable.href) {
+        if (loanUnavailable.href.indexOf('loan/RES') !== -1) {
+          status = 'ausleihbar';
+        } else { status = 'nicht ausleihbar'; }
+      } else {
+        // if there is no url then it will be a presentation
+        if (this.getBookUrl(item) == null) {
+          if (item.label && item.label.indexOf('bestellt') !== -1) {
+            status = item.label;
+            statusInfo = '';
+          } else {
+            status = 'Präsenzbestand';
+            if (presentationAvailable.limitation) {
+              statusInfo = presentationAvailable.limitation[0].content;
+            }
+          }
+        } else {
+          status = 'Online-Ressource im Browser öffnen';
+        }
+      }
+
+      if (presentationUnavailable) {
+        if (loanUnavailable && loanUnavailable.href) {
+          if (loanUnavailable.href.indexOf('loan/RES') !== -1) {
+            status = 'ausgeliehen';
+            if (!loanUnavailable.expected || loanUnavailable.expected === 'unknown') {
+              statusInfo = statusInfo + 'ausgeliehen, Vormerken möglich';
+            } else {
+              statusInfo = statusInfo + 'ausgeliehen bis ';
+              statusInfo = statusInfo + moment(loanUnavailable.expected, 'YYYY-MM-DD').format('DD.MM.YYYY');
+              statusInfo = statusInfo + ', vormerken möglich';
+            }
+          }
+        } else {
+          statusInfo = statusInfo + '...';
+        }
+      }
+    }
+
+    return [status, statusInfo];
+  }
+
+  /**
+   * @name getBookUrl
+   * @param item
+   */
+  getBookUrl(item): void {
+    let url;
+    if (this.bookLocation) {
+      let i;
+      const tmp = utils.convertToArray(this.bookLocation);
+      for (i = 0; i < tmp.length; i++) {
+        if (tmp[i].url) {
+          const tmpUrl = utils.convertToArray(tmp[i].url);
+          let j;
+          for (j = 0; j < tmpUrl.length; j++) {
+            if (tmpUrl[j].$ && (tmpUrl[j].$.usage === 'primary display')) {
+              if (tmpUrl[j]._) {
+                url = tmpUrl[j]._;
+                break;
+              }
+            }
+          }
+        }
+      }
+    }
+
+    if (this.mediaType === 'mediatype_o') {
+      let tmp;
+      if (item.unavailable && item.unavailable[0].service === 'openaccess') {
+        tmp = item.unavailable[0].href;
+      } else { tmp = null; }
+      if (tmp != null) { url = tmp; }
+    }
+
+    return url;
+  }
+
+}
